requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
pydeck>=0.9.1
google-cloud-firestore>=2.11.1
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
protobuf>=4.21.0
